+++ term_ui.h
@@ -11,5 +11,5 @@
 #define NEWLINE_SPACE 1
 #define TAB_SPACE 4
 
-void initialize_ui(char* base, DoublyLinkedList results, int size, int query_len);
+void initialize_ui(char* base, DoublyLinkedList results, uint64_t size, int query_len);
